fonts: Add more types to `fonts_traits`
This is necessary so that `constellation_traits` can get these types via
dependency on `fonts_traits`. This will allow IPC channels to shared
workers so that they can have access to a shared `FontContext` from
`script`.
